djangox/requirements.txt
alabaster==0.7.16
asgiref==3.8.1
attrs==24.2.0
babel==2.16.0
backcall==0.2.0
bleach==6.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Django==5.1
django-admin-view-permission==1.9
django-cors-headers==4.4.0
django-extensions==3.2.3
djangorestframework==3.15.2
docutils==0.21.2
entrypoints==0.4
fastjsonschema==2.20.0
fuzzywuzzy==0.18.0
idna==3.7
imagesize==1.4.1
ipykernel==6.4.2
ipython==7.34.0
ipython-genutils==0.2.0
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-client==7.3.4
jupyter_core==5.0.0
jupyterlab_pygments==0.3.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mistune==0.8.4
more-itertools==10.4.0
nbclient==0.5.11
nbconvert==6.0.7
nbformat==5.3.0
nest-asyncio==1.6.0
notebook==6.0.3
numpy==2.1.0
opencv-python==4.10.0.84
packaging==24.1
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prometheus_client==0.20.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
ptyprocess==0.7.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
pyzmq==26.1.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
Send2Trash==1.8.3
shapely==2.0.6
simplegeneric==0.8.1
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.9.1
Sphinx==2.0.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.5.1
terminado==0.8.2
testpath==0.4.2
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.3
typing==3.7.4.1
typing_extensions==4.12.2
urllib3==1.24.1
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.1
widgetsnbextension==3.4.2
zipp==0.6.0
